requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.2.0
alabaster==0.7.16
alpha_vantage==3.0.0
arrow==1.3.0
astroid==3.3.8
asttokens==2.0.5
asyncssh==2.17.0
atomicwrites==1.4.0
attrs==24.3.0
autopep8==2.0.4
babel==2.16.0
beautifulsoup4==4.12.3
binaryornot==0.4.4
black==24.10.0
bleach==6.2.0
Bottleneck==1.4.2
Brotli==1.0.9
certifi==2025.1.31
cffi==1.17.1
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
comm==0.2.1
contourpy==1.3.1
cookiecutter==2.6.0
cryptography==43.0.3
cycler==0.11.0
debugpy==1.8.11
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
diff-match-patch==20200713
dill==0.3.8
docstring-to-markdown==0.11
docutils==0.18.1
et-xmlfile==1.1.0
executing==0.8.3
fastjsonschema==2.20.0
flake8==7.1.1
fonttools==4.55.3
frozenlist==1.5.0
idna==3.7
imagesize==1.4.1
importlib_metadata==8.5.0
inflection==0.5.1
intervaltree==3.1.0
ipykernel==6.29.5
ipython==8.30.0
isort==5.13.2
jaraco.classes==3.2.1
jedi==0.19.2
jeepney==0.7.1
jellyfish==1.1.3
Jinja2==3.1.5
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.7.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab-pygments==0.2.2
keyring==24.3.1
kiwisolver==1.4.8
lxml==5.3.0
markdown-it-py==2.2.0
MarkupSafe==3.0.2
matplotlib==3.9.1
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.0
mistune==2.0.4
mkl_fft==1.3.11
mkl_random==1.2.8
mkl-service==2.4.0
more-itertools==10.3.0
multidict==6.1.0
mypy-extensions==1.0.0
nbclient==0.8.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
numexpr==2.10.1
numpy==2.0.1
numpydoc==1.7.0
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.0
parso==0.8.4
pathspec==0.10.3
pexpect==4.8.0
pickleshare==0.7.5
pillow==11.1.0
pip==25.0
platformdirs==3.10.0
pluggy==1.5.0
ply==3.11
prompt-toolkit==3.0.43
propcache==0.2.0
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.12.1
pycparser==2.21
pydocstyle==6.3.0
pyflakes==3.2.0
PyGithub==2.4.0
Pygments==2.15.1
PyJWT==2.10.1
pylint==3.3.3
pylint-venv==3.0.3
pyls-spyder==0.4.0
PyNaCl==1.5.0
pyparsing==3.2.0
PyQt5==5.15.10
PyQt5_sip==12.13.0
PyQtWebEngine==5.15.6
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-lsp-black==2.0.0
python-lsp-jsonrpc==1.1.2
python-lsp-server==1.12.0
python-slugify==5.0.2
pytoolconfig==1.2.6
pytz==2024.1
pyuca==1.2
pyxdg==0.27
PyYAML==6.0.2
pyzmq==26.2.0
QDarkStyle==3.2.3
qstylizer==0.2.2
QtAwesome==1.3.1
qtconsole==5.6.1
QtPy==2.4.1
referencing==0.30.2
requests==2.32.3
rich==13.9.4
rope==1.12.0
rpds-py==0.22.3
Rtree==1.0.1
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
SecretStorage==3.3.1
setuptools==75.8.0
sip==6.7.12
six==1.16.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.5
Sphinx==7.3.7
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.10
spyder==6.0.3
spyder-kernels==3.0.2
stack-data==0.2.0
superqt==0.6.7
tabulate==0.9.0
text-unidecode==1.3
textdistance==4.6.3
threadpoolctl==3.5.0
three-merge==0.1.1
tinycss2==1.4.0
tomli==2.0.1
tomlkit==0.13.2
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2023.3
ujson==5.10.0
unicodedata2==15.1.0
Unidecode==1.3.8
urllib3==2.3.0
watchdog==4.0.1
wcwidth==0.2.5
webencodings==0.5.1
whatthepatch==1.0.2
wheel==0.45.1
wrapt==1.17.0
wurlitzer==3.0.2
xgboost==2.1.4
yapf==0.40.2
yarl==1.18.0
zipp==3.21.0
